PGPRO-2033: a try to fix the use of dispell_reset in shared_ispell on Windows
The shared_ispell dictionary is loaded into memory if the timestamp of its last
lookup is less than the timestamp of the reset of the shared dictionary memory.
These timestamps used the function gettimeofday() which is not for Win32 high
precision timing purposes. Therefore sometimes the timestamp of the last
dictinary lookup and the timestamp of the reset of the shared dictionary memory
are equal and the function ts_lexize returns NULL (from buildfarm). To avoid
this use the structure  instr_time and its macros/functions with portable
high-precision interval timing for the same purpose.